Only process new files if they've changed
Using crc32 to determine if a file has changed.
Changed fileId to a 64-bit integer.
Fixed bugs in the integration tests.
Removed a slow query from the status API.

+++ server/service/api_searchpage.go
@@ -15,7 +15,7 @@ type apiMethodSearchPage struct {
 }
 
 type apiSearchPageHit struct {
-	FileID        int        `json:"fileId"`
+	FileID        int64      `json:"fileId"`
 	Filename      jsonString `json:"filename"`
 	IsCommand     bool       `json:"isCommand"`
 	Excerpt       string     `json:"excerpt"`
